requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0

# LLM & Multi-Modal
langchain==0.1.0
langgraph==0.0.20
openai==1.3.0

# Computer Vision - Use newer versions for Windows
opencv-python>=4.8.0
pillow>=10.0.0  # Let pip choose compatible version
paddleocr>=2.7.0
paddlepaddle>=2.5.0  # CPU version

# Document Processing
PyPDF2>=3.0.0
pdfplumber>=0.10.0
pdf2image>=1.16.0
python-docx>=1.0.0
python-pptx>=0.6.0

# Vector Database
qdrant-client>=1.7.0
sentence-transformers>=2.2.0

# Utilities
requests>=2.31.0
numpy>=1.24.0,<2.0.0  # Compatibility constraint
pandas>=2.0.0
tqdm>=4.66.0

# Testing & Quality
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0

# UI (Optional for demo)
streamlit>=1.28.0
gradio>=4.0.0